refactor and added designs

+++ database/migrations/2025_03_01_125141_add_rate_column_to_users.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('rate')->default(1);
+            $table->float('rate')->default(1);
         });
     }
